perf(lysis): widen NodeId/EffectId/SsaVar to 64-bit indices
The lysis emission counter exhausts at u32::MAX = 4.29B emissions, which
ECDSAVerify-class circuits hit during execute. Widen the three coupled
index types end-to-end so multi-billion-emission programs no longer
panic on `node index + 1 overflows u32`.

* NodeId: NonZeroU32 -> NonZeroU64 (lysis-types). The Option<NodeId>
  niche stays single-word post-widening.
* EffectId: NonZeroU32 -> NonZeroU64 (lysis intern).
* SsaVar: u32 -> u64 (ir-core newtype) with IrProgram::next_var and
  getters/setters widening accordingly.
* The lysis-to-R1CS bridge cast widens to `as u64` so the bridge no
  longer truncates on >=2^32 ids.
* Cascade through every site that names NodeId.raw(), SsaVar.0, or
  the bridge cast (ir/, ir-forge/lysis_lift, instantiate, zkc).

Pinned by a co-located unit test (fresh_node_id_advances_past_u32_max)
that sets the NodeIdGen counter just below u32::MAX and verifies fresh()
advances past the boundary without panicking.

The ecdsa_verify_boss_fight integration probe no longer panics on
NonZeroU32 exhaustion. A separate RSS pressure wall surfaces inside
execute under sandboxed runs (~10 GB peak, ~1.1 GB single allocation
fails); the test's #[ignore] reason is rewritten to describe that
present-truth failure mode and the residual cause (per-program state
preserved across chunk-seal boundaries), while folding the closed
NodeId-index-width wall into the existing list of pinned closures.

+++ ir-core/src/types.rs
@@ -5,7 +5,7 @@ use memory::{Bn254Fr, FieldBackend, FieldElement};
 
 /// An SSA variable — defined exactly once.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct SsaVar(pub u32);
+pub struct SsaVar(pub u64);
 
 /// Whether a circuit input is public (instance) or private (witness).
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -517,7 +517,7 @@ impl std::fmt::Display for IrType {
 #[derive(Debug)]
 pub struct IrProgram<F: FieldBackend = Bn254Fr> {
     pub instructions: Vec<Instruction<F>>,
-    pub next_var: u32,
+    pub next_var: u64,
     pub var_names: HashMap<SsaVar, String>,
     pub var_types: HashMap<SsaVar, IrType>,
     pub input_spans: HashMap<String, SpanRange>,
@@ -649,14 +649,14 @@ impl<F: FieldBackend> IrProgram<F> {
     }
 
     /// Current `next_var` watermark (the id the next `fresh_var()` will return).
-    pub fn next_var(&self) -> u32 {
+    pub fn next_var(&self) -> u64 {
         self.next_var
     }
 
     /// Force the `next_var` watermark — needed by passes that re-number SSA
     /// (canonicalization, oracle harness setup). Avoid in normal compile paths;
     /// use `fresh_var()` instead.
-    pub fn set_next_var(&mut self, n: u32) {
+    pub fn set_next_var(&mut self, n: u64) {
         self.next_var = n;
     }
